requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
supabase==1.0.3
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-dotenv==1.0.0
python-multipart==0.0.5
pydantic==1.10.12